MOHAWK: Fix the Riven demo intro sequence (a regression from r55188)
svn-id: r55484

+++ engines/mohawk/video.h
@@ -57,7 +57,7 @@ struct VideoEntry {
 
 	// Identification
 	Common::String filename; // External video files
-	uint16 id;               // Internal Mohawk files
+	int id;                  // Internal Mohawk files
 
 	// Helper functions
 	Video::SeekableVideoDecoder *operator->() const { assert(video); return video; } // TODO: Remove this eventually
